scope module
